Moved App.jsx to Components folder and updated BoDChalkboard styles
